+++ src/titleui/old.cpp
@@ -102,9 +102,6 @@ void WzOldTitleUI::start()
 	case MULTI:
 		startMultiPlayerMenu();		// goto multiplayer menu
 		break;
-	case GAMEFIND:
-		startGameFind();
-		break;
 	case KEYMAP:
 		startKeyMapEditor(true);
 		break;
@@ -133,9 +130,6 @@ TITLECODE WzOldTitleUI::run()
 	switch (mode) // run relevant title screen code.
 	{
 	// MULTIPLAYER screens
-	case GAMEFIND:
-		runGameFind();
-		break;
 	case MULTI:
 		runMultiPlayerMenu();
 		break;
